data_val/requirements.txt
pdfplumber
ruamel.yaml
google-generativeai
pytest
PyYAML 